+++ widgets/EntryParticipants.php
@@ -61,9 +61,16 @@ class EntryParticipants extends Widget
         }
 
         $participantSate = $calendarEntry->getParticipationStatus(Yii::$app->user->identity);
-        return Button::info($label)
-                     ->icon($participantSate === $state ? 'fa-check-circle' : null)
-                     ->cssClass($participantSate === $state ? '' : 'active')
-                     ->action('calendar.respond', Url::toEntryRespond($calendarEntry, $state));
-    }
+
+        $button = Button::info($label)
+            ->icon($participantSate === $state ? 'fa-check-circle' : null);
+        if ($calendarEntry->isPast()) {
+            $button->cssClass('disabled')->loader(false);
+        } else {
+            $button->action('calendar.respond', Url::toEntryRespond($calendarEntry, $state))
+                ->cssClass($participantSate === $state ? '' : 'active');
+        }
+
+        return $button;
+   }
 }